Time: 9 ms (44.18%), Space: 14.4 MB (10.01%) - LeetHub

+++ 0260-single-number-iii/0260-single-number-iii.cpp
@@ -1,7 +1,7 @@
 class Solution {
 public:
     vector<int> singleNumber(vector<int>& nums) {
-        unordered_map<int,int>mp;
+        unordered_map<int,short>mp;
         vector<int> sol;
         for(auto it:nums){
             mp[it]++;
